+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.brick",
+    package = "com.yourcare",
 )
 
 android_resource(
     name = "res",
-    package = "com.brick",
+    package = "com.yourcare",
     res = "src/main/res",
 )
