+++ android/app/src/main/java/com/yourcare/MainActivity.java
@@ -1,4 +1,4 @@
-package com.brick;
+package com.yourcare;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -12,7 +12,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "brick";
+    return "yourcare";
   }
 
   /**
