+++ android/app/src/main/java/com/yourcare/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.brick.newarchitecture.components;
+package com.yourcare.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
