+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "brick",
-  "displayName": "brick"
+  "name": "yourcare",
+  "displayName": "yourcare"
 }
