+++ ios/yourcare.xcodeproj/xcshareddata/xcschemes/yourcare.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "brick.app"
-               BlueprintName = "brick"
-               ReferencedContainer = "container:brick.xcodeproj">
+               BuildableName = "yourcare.app"
+               BlueprintName = "yourcare"
+               ReferencedContainer = "container:yourcare.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "brickTests.xctest"
-               BlueprintName = "brickTests"
-               ReferencedContainer = "container:brick.xcodeproj">
+               BuildableName = "yourcareTests.xctest"
+               BlueprintName = "yourcareTests"
+               ReferencedContainer = "container:yourcare.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "brick.app"
-            BlueprintName = "brick"
-            ReferencedContainer = "container:brick.xcodeproj">
+            BuildableName = "yourcare.app"
+            BlueprintName = "yourcare"
+            ReferencedContainer = "container:yourcare.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "brick.app"
-            BlueprintName = "brick"
-            ReferencedContainer = "container:brick.xcodeproj">
+            BuildableName = "yourcare.app"
+            BlueprintName = "yourcare"
+            ReferencedContainer = "container:yourcare.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
