+++ ios/yourcare/Info.plist
@@ -5,7 +5,7 @@
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
 	<key>CFBundleDisplayName</key>
-	<string>brick</string>
+	<string>yourcare</string>
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
@@ -26,10 +26,10 @@
 			<key>CFBundleTypeRole</key>
 			<string>Editor</string>
 			<key>CFBundleURLName</key>
-			<string>mbrick</string>
+			<string>myourcare</string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>mbrick</string>
+				<string>myourcare</string>
 			</array>
 		</dict>
 	</array>
