+++ ios/yourcareTests/yourcareTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface brickTests : XCTestCase
+@interface yourcareTests : XCTestCase
 
 @end
 
-@implementation brickTests
+@implementation yourcareTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
