+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "brick",
+  "name": "yourcare",
   "version": "0.0.1",
   "private": true,
   "scripts": {
