+++ src/components/atoms/index.tsx
@@ -1,0 +1,2 @@
+export * from './BasicButton';
+export * from './BasicInput';
