+++ src/docs/fastlane.md
@@ -1,0 +1,1 @@
+https://millo-l.github.io/ReactNative-fastlane/
