+++ src/docs/firebase_conf.md
@@ -1,0 +1,4 @@
+https://devbksheen.tistory.com/m/entry/React-Native-%ED%94%84%EB%A1%9C%EC%A0%9D%ED%8A%B8-Firebase-%EC%97%B0%EB%8F%99%ED%95%98%EA%B8%B0
+
+type
+https://rnfirebase.io/typescript
