+++ src/docs/rn_tdd.md
@@ -1,0 +1,7 @@
+https://skyksit.tistory.com/entry/RNReactNative-TDD-%EC%84%A4%EC%A0%95
+
+이대로 했다간 개고생
+
+https://mercedesbernard.com/blog/jest-mocking-strategies#in-a-single-file
+
+https://inpa.tistory.com/entry/JEST-%F0%9F%93%9A-%EB%AA%A8%ED%82%B9-mocking-jestfn-jestspyOn#mockfnmockreturnvaluevalue
