+++ src/docs/typescript.md
@@ -1,0 +1,1 @@
+https://soopdop.github.io/2020/12/01/index-signatures-in-typescript/
